feat: reservations, validations and business logic

+++ APBD-PJATK-Cw3/Services/IRoomsService.cs
@@ -6,9 +6,9 @@ namespace APBD_PJATK_Cw3.Services;
 public interface IRoomsService
 {
     public List<RoomResponse> GetRooms(int? minCapacity, bool? hasProjector, bool? isActive);
-    public RoomResponse GetRoom(int id);
+    public RoomResponse GetRoom(long id);
     public List<RoomResponse> GetByBuildingCode(string buildingCode);
     public RoomResponse AddRoom(AddRoomRequest request);
-    public RoomResponse UpdateRoom(UpdateRoomRequest request, int id);
-    public void DeleteRoom(int id);
+    public RoomResponse UpdateRoom(UpdateRoomRequest request, long id);
+    public void DeleteRoom(long id);
 }
